Fix PHP 8.4 strict typing across Validator, DI, Yaml, and EventDispatcher.
Normalize violation parameters, container parameter bags, reference IDs, and parser ref tracking for PHP 8.4; load ServicesBundle before FrameworkExtension in listener pass tests.

+++ src/Symfony/Component/Validator/Constraints/DateValidator.php
@@ -61,9 +61,9 @@ class DateValidator extends ConstraintValidator
         }
 
         if (!self::checkDate(
-            $matches['year'] ?? $matches[1],
-            $matches['month'] ?? $matches[2],
-            $matches['day'] ?? $matches[3]
+            (int) ($matches['year'] ?? $matches[1]),
+            (int) ($matches['month'] ?? $matches[2]),
+            (int) ($matches['day'] ?? $matches[3])
         )) {
             $this->context->buildViolation($constraint->message)
                 ->setParameter('{{ value }}', $this->formatValue($value))
